+++ smart-contracts/contracts-common/concordium-contracts-common/src/types.rs
@@ -492,7 +492,6 @@ impl str::FromStr for Timestamp {
     type Err = ParseTimestampError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        use convert::TryInto;
         let datetime =
             chrono::DateTime::parse_from_rfc3339(s).map_err(ParseTimestampError::ParseError)?;
         let millis = datetime
@@ -1765,6 +1764,30 @@ repeat_macro!(
 /// since the values are easier to construct.
 pub type OwnedPolicy = Policy<Vec<(AttributeTag, AttributeValue)>>;
 
+impl OwnedPolicy {
+    /// Serialize the policy for consumption by smart contract execution engine.
+    ///
+    /// This entails the following serialization scheme:
+    /// - `1`:             u8            specifying a single policy.
+    /// - `len`:           u16           length of the inner payload
+    /// - `inner payload`: `len` bytes   the serialized `OwnedPolicy`
+    #[doc(hidden)]
+    pub fn serial_for_smart_contract<W: crate::traits::Write>(
+        &self,
+        out: &mut W,
+    ) -> Result<(), W::Err> {
+        // Serialize to an inner vector.
+        let inner = to_bytes(self);
+        // Specify that there is only one policy.
+        out.write_u8(1)?;
+        // Write length of the inner.
+        (inner.len() as u16).serial(out)?;
+        // Write the inner buffer.
+        out.write_all(&inner)?;
+        Ok(())
+    }
+}
+
 /// Index of the identity provider on the chain.
 /// An identity provider with the given index will not be replaced,
 /// so this is a stable identifier.
@@ -2114,8 +2137,6 @@ mod serde_impl {
         type Err = ExchangeRateConversionError;
 
         fn from_str(s: &str) -> Result<Self, Self::Err> {
-            use convert::TryInto;
-
             let mut decimal = rust_decimal::Decimal::from_str_exact(s)?;
             decimal.normalize_assign();
             if decimal.is_zero() || decimal.is_sign_negative() {